+++ lib/brisket.js
@@ -28,8 +28,6 @@ var Brisket = {
         StringTemplateAdapter: require("./templating/StringTemplateAdapter")
     },
 
-    Testing: require("./testing/Testing"),
-
     Backbone: require("./application/Backbone"),
 
     $: require("./application/jquery"),
