+++ lib/server/Server.js
@@ -26,13 +26,6 @@ var Server = {
         var redirectOnNewLayout = config.redirectOnNewLayout === true;
 
         Deprecated.message(
-            "Accessing Brisket.Testing from require('brisket').Testing is " +
-            "deprecated. require('brisket').Testing will be removed in the next " +
-            "major release. Use require('brisket/testing') instead",
-            "https://github.com/bloomberg/brisket/blob/master/docs/brisket.testing.md"
-        );
-
-        Deprecated.message(
             "Brisket.createServer({ redirectOnNewLayout: false }) is deprecated. " +
             "Brisket will always do a full redirect when a route uses a different " +
             "Layout in the next major release."
