--- spec/server/brisketTestingSpec.js
@@ -1,28 +1,0 @@
-"use strict";
-
-describe("public interface to Brisket Testing", function() {
-    var BrisketTesting = require("../../testing");
-
-    it("exposes brisket/testing", function() {
-        expect(BrisketTesting).toBe(requireFromLib("testing/Testing"));
-    });
-
-    function requireFromLib(path) {
-        return require("../../lib/" + path);
-    }
-});
-
-// ----------------------------------------------------------------------------
-// Copyright (C) 2017 Bloomberg Finance L.P.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// ----------------------------- END-OF-FILE ----------------------------------
